Share CLI HTTP helpers
