+++ lmcache/cli/commands/describe.py
@@ -8,13 +8,16 @@ Usage::
 
 # Standard
 import argparse
-import json
 import sys
-import urllib.error
-import urllib.request
 
 # First Party
 from lmcache.cli.commands.base import BaseCommand
+from lmcache.cli.http import (
+    CLIHTTPError,
+    DEFAULT_URLS,
+    fetch_json as _fetch_json,
+    normalize_url,
+)
 from lmcache.cli.metrics import Metrics
 
 # -------------------------------------------------------------------
@@ -26,36 +29,16 @@ class DescribeError(Exception):
     """Raised when the describe command cannot fetch or parse status data."""
 
 
-def normalize_url(url: str) -> str:
-    """Ensure *url* has an ``http://`` or ``https://`` scheme."""
-    if not url.startswith(("http://", "https://")):
-        url = f"http://{url}"
-    return url.rstrip("/")
-
-
 def fetch_json(url: str, timeout: int = 10) -> dict:
     """GET *url* and return the parsed JSON body.
 
     Raises:
         DescribeError: On network/HTTP errors.
     """
-    req = urllib.request.Request(url)
     try:
-        with urllib.request.urlopen(req, timeout=timeout) as resp:
-            return json.loads(resp.read().decode())
-    except urllib.error.HTTPError as exc:
-        if exc.code == 503:
-            body = exc.read().decode()
-            try:
-                detail = json.loads(body).get("error", body)
-            except (json.JSONDecodeError, AttributeError):
-                detail = body
-            raise DescribeError(f"Server unhealthy: {detail}") from exc
-        raise DescribeError(f"HTTP {exc.code} from {url}: {exc.reason}") from exc
-    except urllib.error.URLError as exc:
-        raise DescribeError(f"Cannot connect to {url}: {exc.reason}") from exc
-    except OSError as exc:
-        raise DescribeError(f"Cannot connect to {url}: {exc}") from exc
+        return _fetch_json(url, timeout=timeout)
+    except CLIHTTPError as exc:
+        raise DescribeError(str(exc)) from exc
 
 
 def fmt_bytes(n: int) -> str:
@@ -321,7 +304,7 @@ class DescribeCommand(BaseCommand):
         parser.add_argument(
             "--url",
             help="LMCache HTTP server URL (default to http://localhost:8080).",
-            default="http://localhost:8080",
+            default=DEFAULT_URLS["kvcache"],
         )
 
     def execute(self, args: argparse.Namespace) -> None:
