+++ lmcache/cli/http.py
@@ -1,0 +1,49 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Shared HTTP helpers for CLI commands."""
+
+# Standard
+from typing import Any
+import json
+import urllib.error
+import urllib.request
+
+DEFAULT_URLS: dict[str, str] = {
+    "kvcache": "http://localhost:8080",
+    "engine": "http://localhost:8000",
+}
+
+
+class CLIHTTPError(Exception):
+    """Raised when a CLI HTTP request fails."""
+
+
+def normalize_url(url: str) -> str:
+    """Ensure *url* has an ``http://`` or ``https://`` scheme."""
+    if not url.startswith(("http://", "https://")):
+        url = f"http://{url}"
+    return url.rstrip("/")
+
+
+def fetch_json(url: str, timeout: int = 10) -> dict[str, Any]:
+    """GET *url* and return the parsed JSON body.
+
+    Raises:
+        CLIHTTPError: On network/HTTP errors.
+    """
+    req = urllib.request.Request(url)
+    try:
+        with urllib.request.urlopen(req, timeout=timeout) as resp:
+            return json.loads(resp.read().decode())
+    except urllib.error.HTTPError as exc:
+        if exc.code == 503:
+            body = exc.read().decode()
+            try:
+                detail = json.loads(body).get("error", body)
+            except (json.JSONDecodeError, AttributeError):
+                detail = body
+            raise CLIHTTPError(f"Server unhealthy: {detail}") from exc
+        raise CLIHTTPError(f"HTTP {exc.code} from {url}: {exc.reason}") from exc
+    except urllib.error.URLError as exc:
+        raise CLIHTTPError(f"Cannot connect to {url}: {exc.reason}") from exc
+    except OSError as exc:
+        raise CLIHTTPError(f"Cannot connect to {url}: {exc}") from exc
